ElearningProject/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==4.0.1
Django==4.1.9
django-allauth==0.54.0
django-cors-headers==4.0.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.1.3
ecdsa==0.18.0
idna==3.4
inflection==0.5.1
jwcrypto==1.5.0
oauth2-provider==0.0
oauthlib==3.2.2
Pillow==9.5.0
psycopg==3.1.9
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
wrapt==1.15.0
